增加新建用户名下车辆时的检查，修改用户名下车辆管理的删除方法。

+++ src/main/java/io/renren/modules/business/entity/CustomerCarSeriesRelEntity.java
@@ -24,15 +24,15 @@ public class CustomerCarSeriesRelEntity implements Serializable {
 	 * 关联id
 	 */
 	@TableId
-	private Integer relId;
+	private Long relId;
 	/**
 	 * 客户id
 	 */
-	private Integer customerId;
+	private Long customerId;
 	/**
 	 * 车系id
 	 */
-	private Integer seriesId;
+	private Long seriesId;
 	/**
 	 * 车牌号码
 	 */
